refactor: move some method from write script
